backend/requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.1
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.12
decorator==5.1.1
distro==1.9.0
Django==5.1.5
dnspython==2.7.0
docx2txt==0.9
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==5.29.1
gradio_client==1.10.1
groovy==0.1.2
groq==0.28.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.4.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
loguru==0.7.3
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.13
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
openai==1.93.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
preshed==3.0.10
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
routers==0.10.1
rpds-py==0.25.1
rsa==4.9.1
ruff==0.11.10
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.1
selenium==4.30.0
semantic-version==2.10.0
sentence-transformers==5.0.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
# Editable Git install with no remote (startup_investor==0.1)
-e /Users/jaswanthb/Desktop/startup_investor_platform/backend
streamlit==1.46.0
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.5.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchinfo==1.8.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2024.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
wsproto==1.2.0
xlsxwriter==3.2.5
